requirements.txt
gspread==5.12.2
google-auth==2.29.0
supabase==2.4.0
pandas==2.2.2